修复: 消除所有 float64→complex128 路径, 统一使用 float32→complex64 (NPU 兼容)
- model.py sinusoidal_embedding_1d: float64→float32
- model.py sigma_theta build_freqs: 3处 float64→float32 (linspace/arange pos/exp)
- action_module.py sinusoidal_embedding_1d: float64→float32
- posemb_layers.py 已使用 float32, 无需修改

+++ Matrix-Game-3/wan/modules/action_module.py
@@ -32,7 +32,7 @@ class WanRMSNorm(nn.Module):
 def sinusoidal_embedding_1d(dim, position):
     assert dim % 2 == 0
     half = dim // 2
-    position = position.type(torch.float64)
+    position = position.float()  # NPU: float32 避免 double 算子不兼容
 
     sinusoid = torch.outer(position, torch.pow(10000, -torch.arange(half).to(position).div(half)))
     x = torch.cat([torch.cos(sinusoid), torch.sin(sinusoid)], dim=1)
